renamed controller file to a clearer name
